Move Elasticsearchserver to tests.es package
It is only used for tests so this is a more natural place for it. But this is mainly done in preparation to enabling info level logs for Elasticsearchserver in the next commit.
